Refactor PersonalizationTagManager registering tags
